Used a Gaussian approximation for the delta function instead of Lorenztian
So, I used a Gaussian approximation instead of a Lorentzian like Aiden did for the delta function. I was able to reduce the ratio of the powers from around 5.56 to about 3.12. Also, the power from the distribution is now around 1.06 instead of 1.88 indicating that the Gaussian was a better choice.